functions/auth-get-user/requirements.txt
yandexcloud==0.328.0
ydb[yc]==3.8.1
PyJWT==2.8.0
bcrypt==4.1.2 